refactor : 인앱 업데이트를 MainActivity에서 수행하도록 변경
