+++ app/src/main/java/online/partyrun/partyrunapplication/AuthActivity.kt
@@ -2,26 +2,11 @@ package online.partyrun.partyrunapplication
 
 import android.content.Intent
 import android.os.Bundle
-import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.lifecycle.lifecycleScope
-import com.google.android.play.core.appupdate.AppUpdateManager
-import com.google.android.play.core.appupdate.AppUpdateManagerFactory
-import com.google.android.play.core.appupdate.AppUpdateOptions
-import com.google.android.play.core.install.InstallStateUpdatedListener
-import com.google.android.play.core.install.model.ActivityResult
-import com.google.android.play.core.install.model.AppUpdateType
-import com.google.android.play.core.install.model.InstallStatus
-import com.google.android.play.core.install.model.UpdateAvailability
-import com.google.android.play.core.ktx.isFlexibleUpdateAllowed
-import com.google.android.play.core.ktx.isImmediateUpdateAllowed
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 import online.partyrun.partyrunapplication.core.common.extension.setIntentActivity
 import online.partyrun.partyrunapplication.core.designsystem.theme.PartyRunApplicationTheme
 import online.partyrun.partyrunapplication.ui.PartyRunAuth
@@ -29,21 +14,9 @@ import online.partyrun.partyrunapplication.ui.PartyRunAuth
 @AndroidEntryPoint
 class AuthActivity : ComponentActivity() {
 
-    private lateinit var appUpdateManager: AppUpdateManager
-
-    // 업데이트 타입 변경 가능 -> 현재 FLEXIBLE
-    private val updateType = AppUpdateType.FLEXIBLE
-
     override fun onCreate(savedInstanceState: Bundle?) {
         // Splash Screen API
         installSplashScreen()
-
-        // 최신 업데이트 정보 확인
-        appUpdateManager = AppUpdateManagerFactory.create(applicationContext)
-        if (updateType == AppUpdateType.FLEXIBLE) {
-            appUpdateManager.registerListener(installStateUpdatedListener)
-        }
-        checkForAppUpdates()
 
         super.onCreate(savedInstanceState)
         setContent {
@@ -68,83 +41,6 @@ class AuthActivity : ComponentActivity() {
         )
         overridePendingTransition(R.anim.slide_in_right, R.anim.slide_out_left) // 화면 전환 애니메이션
         finish()
-    }
-
-    private val activityResultLauncher =
-        registerForActivityResult(ActivityResultContracts.StartIntentSenderForResult()) { result ->
-            when (result.resultCode) {
-                ActivityResult.RESULT_IN_APP_UPDATE_FAILED ->
-                    Toast.makeText(
-                        applicationContext,
-                        applicationContext.getString(R.string.result_in_app_update_failed),
-                        Toast.LENGTH_SHORT
-                    ).show()
-            }
-        }
-
-    private fun checkForAppUpdates() {
-        val appUpdateInfoTask = appUpdateManager.appUpdateInfo
-
-        appUpdateInfoTask.addOnSuccessListener { info ->
-            val isUpdateAvailable = info.updateAvailability() == UpdateAvailability.UPDATE_AVAILABLE
-            val isUpdateAllowed = when (updateType) {
-                AppUpdateType.IMMEDIATE -> info.isImmediateUpdateAllowed
-                AppUpdateType.FLEXIBLE -> info.isFlexibleUpdateAllowed
-                else -> false
-            }
-            if (isUpdateAvailable && isUpdateAllowed) {
-                appUpdateManager.startUpdateFlowForResult(
-                    info,
-                    activityResultLauncher,
-                    AppUpdateOptions.newBuilder(AppUpdateType.FLEXIBLE)
-                        .setAllowAssetPackDeletion(true)
-                        .build()
-                )
-            }
-        }
-    }
-
-    private val installStateUpdatedListener = InstallStateUpdatedListener { state ->
-        if (state.installStatus() == InstallStatus.DOWNLOADED) {
-            val msg = applicationContext.getString(R.string.update_download_success)
-            Toast.makeText(
-                applicationContext,
-                msg,
-                Toast.LENGTH_LONG
-            ).show()
-            lifecycleScope.launch {
-                delay(3000)
-                appUpdateManager.completeUpdate()
-            }
-        }
-    }
-
-    override fun onResume() {
-        super.onResume()
-        if (updateType == AppUpdateType.IMMEDIATE) {
-            val appUpdateInfoTask = appUpdateManager.appUpdateInfo
-
-            appUpdateInfoTask.addOnSuccessListener { info ->
-                val isUpdateAvailable =
-                    info.updateAvailability() == UpdateAvailability.DEVELOPER_TRIGGERED_UPDATE_IN_PROGRESS
-                if (isUpdateAvailable) {
-                    appUpdateManager.startUpdateFlowForResult(
-                        info,
-                        activityResultLauncher,
-                        AppUpdateOptions.newBuilder(AppUpdateType.IMMEDIATE)
-                            .setAllowAssetPackDeletion(true)
-                            .build()
-                    )
-                }
-            }
-        }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        if (updateType == AppUpdateType.FLEXIBLE) {
-            appUpdateManager.unregisterListener(installStateUpdatedListener)
-        }
     }
 
 }
