+++ app/src/main/java/online/partyrun/partyrunapplication/MainActivity.kt
@@ -4,6 +4,7 @@ import android.Manifest
 import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Bundle
+import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.result.contract.ActivityResultContracts
@@ -12,7 +13,18 @@ import androidx.annotation.RequiresApi
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.lifecycleScope
+import com.google.android.play.core.appupdate.AppUpdateManager
+import com.google.android.play.core.appupdate.AppUpdateManagerFactory
+import com.google.android.play.core.appupdate.AppUpdateOptions
+import com.google.android.play.core.install.InstallStateUpdatedListener
+import com.google.android.play.core.install.model.ActivityResult
+import com.google.android.play.core.install.model.AppUpdateType
+import com.google.android.play.core.install.model.InstallStatus
+import com.google.android.play.core.install.model.UpdateAvailability
+import com.google.android.play.core.ktx.isFlexibleUpdateAllowed
+import com.google.android.play.core.ktx.isImmediateUpdateAllowed
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import online.partyrun.partyrunapplication.ui.PartyRunMain
 import online.partyrun.partyrunapplication.core.common.extension.setIntentActivity
@@ -24,6 +36,11 @@ import timber.log.Timber
 class MainActivity : ComponentActivity() {
 
     private val battleViewModel: BattleViewModel by viewModels()
+
+    private lateinit var appUpdateManager: AppUpdateManager
+
+    // 업데이트 타입 변경 가능 -> 현재 IMMEDIATE
+    private val updateType = AppUpdateType.IMMEDIATE
 
     /**
      * 사용자로부터 권한을 요청하고 그 결과를 받아 처리하기 위한 코드
@@ -68,6 +85,13 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
+        // 최신 업데이트 정보 확인
+        appUpdateManager = AppUpdateManagerFactory.create(applicationContext)
+        if (updateType == AppUpdateType.FLEXIBLE) {
+            appUpdateManager.registerListener(installStateUpdatedListener)
+        }
+        checkForAppUpdates()
+
         battleViewModel.terminateOngoingBattle() // 앱이 처음 시작될 때 진행 중인 배틀이 있다면 종료 요청
         askPermissions() // 권한 확인 및 요청
 
@@ -88,9 +112,84 @@ class MainActivity : ComponentActivity() {
         }
     }
 
+    private val activityResultLauncher =
+        registerForActivityResult(ActivityResultContracts.StartIntentSenderForResult()) { result ->
+            when (result.resultCode) {
+                ActivityResult.RESULT_IN_APP_UPDATE_FAILED ->
+                    Toast.makeText(
+                        applicationContext,
+                        applicationContext.getString(R.string.result_in_app_update_failed),
+                        Toast.LENGTH_SHORT
+                    ).show()
+            }
+        }
+
+    private fun checkForAppUpdates() {
+        val appUpdateInfoTask = appUpdateManager.appUpdateInfo
+
+        appUpdateInfoTask.addOnSuccessListener { info ->
+            val isUpdateAvailable = info.updateAvailability() == UpdateAvailability.UPDATE_AVAILABLE
+            val isUpdateAllowed = when (updateType) {
+                AppUpdateType.IMMEDIATE -> info.isImmediateUpdateAllowed
+                AppUpdateType.FLEXIBLE -> info.isFlexibleUpdateAllowed
+                else -> false
+            }
+            if (isUpdateAvailable && isUpdateAllowed) {
+                appUpdateManager.startUpdateFlowForResult(
+                    info,
+                    activityResultLauncher,
+                    AppUpdateOptions.newBuilder(updateType)
+                        .build()
+                )
+            }
+        }
+    }
+
+    private val installStateUpdatedListener = InstallStateUpdatedListener { state ->
+        if (state.installStatus() == InstallStatus.DOWNLOADED) {
+            val msg = applicationContext.getString(R.string.update_download_success)
+            Toast.makeText(
+                applicationContext,
+                msg,
+                Toast.LENGTH_LONG
+            ).show()
+            lifecycleScope.launch {
+                delay(3000)
+                appUpdateManager.completeUpdate()
+            }
+        }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        if (updateType == AppUpdateType.IMMEDIATE) {
+            val appUpdateInfoTask = appUpdateManager.appUpdateInfo
+
+            appUpdateInfoTask.addOnSuccessListener { info ->
+                val isUpdateAvailable =
+                    info.updateAvailability() == UpdateAvailability.DEVELOPER_TRIGGERED_UPDATE_IN_PROGRESS
+                if (isUpdateAvailable) {
+                    appUpdateManager.startUpdateFlowForResult(
+                        info,
+                        activityResultLauncher,
+                        AppUpdateOptions.newBuilder(AppUpdateType.IMMEDIATE)
+                            .build()
+                    )
+                }
+            }
+        }
+    }
+
     private fun performSignOutProcess() {
         setIntentActivity(AuthActivity::class.java)
         overridePendingTransition(0, 0) // 전환 애니메이션 생략
         finish()
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        if (updateType == AppUpdateType.FLEXIBLE) {
+            appUpdateManager.unregisterListener(installStateUpdatedListener)
+        }
     }
 }
